Added another enemy and a new WIP bullet type
Changed some access modifiers on bullet and gave it a move function that children can override. Added a turret enemy that tracks very quickly but is stationary. Sine bullets will osciliate as they travel. (currently doesn't work when at 45 degrees in the first or third quadrents relative to the enemy.